reusing provisioned cluster, using multiple certs
It's often more convenient to reuse a cluster once it has been provisioned. Added notes for that.
It's also more possible/likely you don't use the same certificate for both the service principal and the management connection to the cluster. Added notes for that too.
Finally, when reusing the cluster for multiple builds without changing the manifest every time, you need to provide an additional powershell argument to overwrite an existing ApplicationType.