Updated PowerShell install link
Updated the link to the PowerShell installation instructions. This has been simplified from the original link and is an easier installation process overall. I also think that this will be the more accurate document long term to link to, since this seems to be the official instructions that we keep updated.