Added option 'nofail' to fstab
I added the option 'nofail' to the fstab configuration. This option may help to prevent that a VM gets not started in the case the filesystem is corrupt or the disk is not available at boot time. So this avoids a lot of SR's we see in CSS and the customer is in a better position to repair the disk then. Otherwise it is required to follow a process liek described in this blog https://blogs.msdn.microsoft.com/linuxonazure/2016/07/21/cannot-ssh-to-linux-vm-after-adding-data-disk-to-etcfstab-and-rebooting/