M16846: Premium disks types: three or five?
Hello, @neilpeterson,

Translator has reported possible source content issue. Based on the five types highlighted in brackets, the source should read "Premium disks come in five types (P10, P20, P30, P40, P50), the size of the disk determines the disk type."

Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR. 

Many thanks in advance.