Remove usage of single quote
- As discussed in this Stack Overflow question:

https://stackoverflow.com/questions/49131353/powershell-script-is-not-executing-when-using-azurerm-virtual-machine-extension

The usage of single quotes in the commandToExecute section of the JSON causes the script not to be executed, with the error:

`Processing -File ''test.ps1'' failed because the file does not have a '.ps1' extension. Specify a valid Windows PowerShell script file name, and then try again.`

I have changed this section to use an escaped double quote, rather than a single quote.