removed direct link to SUSE document
removed the direct link to the SUSE HANA-SR document, as this is a static link and if document updates happen it will be a new link with is not recognized. Replaced it with suse best practice landing page which has always the link to the up-to-date documents. In adddition it show also the missing Netweaver document.