Use local filesystem in AWS service fabric
The AWS tutorial does not join the machines to an Active Directory, instead electing to use the same Administrator password for all machines.

Service Fabric executables run under service accounts that do not share the same password across different machines. Setting up a file share on a single node is a bad idea because none of the other nodes will have write access to the fileshare. This will cause services that depend on file share access to fail at certain points, such as when updating the cluster.