Updates due to tech review
Per Thomas

Remove the reference: 
For more information on creating virtual machines, see one of the following topics:
?	Quickstart: Create a Windows virtual machine in the Azure portal
?	Quickstart: Create a Linux virtual machine in the Azure portal
This is common knowledge. We do not need to include them.
2.	Remove the following sentence with screenshot. By default the drop down menu will not show. 
Be sure that you select Resource Manager from the deployment model drop-down menu: 
3.	There are format issues in Notes and Azure Resource Manager template script sections.  


For more information on creating virtual machines, see one of the following topics:
?	Quickstart: Create a Windows virtual machine in the Azure portal
?	Quickstart: Create a Linux virtual machine in the Azure portal
This is common knowledge. We do not need to include them.