Udate AzureDiskEncryption settings
The new **`KeyVaultResourceId`** and **`KekVaultResourceId`** are _mandatory_ parameters in 2.2 now, so they REALLY need to be documented here. Of course, it would be great if we had documentation for both the 1.1 and 2.2 schemas, but the values here weren't quite right for either.

I have left the `SequenceVersion` although it doesn't seem to _do_ anything.