changed wording
in section active/active deployment I changed the term "dialog instances" to "application server"....this is the more correct term in the SAP universe. There are customers who configure an application server in that way, that they only process batch jobs (batch workload) on it. For those the term "dialog instance" may be a bit misleading. The term "application server" is more common!

Thanks for any comments or questions - cheers Frank