VM needs to be deallocated
Rather than calling out that a VM will restart, we should warn that the VM must be deallocated to resize the OS disk.  (similar to a restart, but the steps of the article cannot be completed without deallocating the VM.  Step 4 of the procedure even says to stop the VM before resizing.