Only set service startup type if not the default to avoid Access Denied errors
In my testing (at least on Windows 10 RS3), setting the startup type on some services (like `bfe` and `dhcp`) gives an Access Denied error. However, these services are managed by Windows and also should already be in the correct state in these cases.

Update `prepare-for-upload-vhd-image.md` so that it only tries to change the service startup type if it isn't already the correct default.