Removed table for maximum disks
A use pointed out in GitHub issue 49160 that having the table list out the maximum disks was redundant and confusing because ever VM type with enough CPUs can have an absolute max of 64 disks.  So I removed the table and incorporated what it was trying to show into the text directly.