Replace no-break space in disks-upload-vhd-to-managed-disk-powershell.md
There is a Unicode character 'NARROW NO-BREAK SPACE' (U+202F) in the sample usage of AzCopy.exe.
AzCopy.exe prints a strange error in some environment (I saw the error in PowerShell 5.1 on Windows 10 build 19603, but it works in PowerShell 7.0). Changing the space to a regular space can fix this issue.