Update service-fabric-patterns-networking.md
Removed the duplicate parameter "virtualNetworkName" so that it won't prompt to enter the VNET name again in cluster deployment blade in Azure Portal.
This parameter is not being used by the template as we have added another parameter for existing VNET Name. This additional VNET parameter needs to be removed.