Included Azure CLI script to install LAD 3.0 extension for VMSS instance
I found it very difficult to find the document to install LAD 3.0 extension for VMSS instance because all of the document talked about VM only. Even customer felt the same thing. So I took time to modify update the script to install LAD 3.0 extension for VMSS instance as well as this will empower our end users to do it themselves.