Update disk-encryption-isolated-network.md
From support:  "The package grub-pc-bin is a dependency for package grub-pc, which is used for installing grub when using the command grub-install. When installed, the package creates a directory called i386-pc under /usr/lib/grub, that is required when you run grub-install without the –target flag, by default the process looks for a file under /usr/lib/grub/i386-pc. In Ubuntu 18.04 images when running apt-upgrade there is a message saying that package grub-pc-bin is not required any longer: