Add output json to cli with grep
The az cli can be configured to return non-json format by invoking `az configure`. In my case, I configured the output to be a table, as this is more readable to me.

```txt
# az configure
Your settings can be found at /home/<User>/.azure/config

[core]
output = table
```

In my opinion, we should not assume the output to be JSON for everyone, but rather be explicit when using a follow-up like `grep` as it is case-sensitive. 
Given the changed code it will always return a value, no matter the default configuration of az,