Remove extra unnecessary backtick characeters
The code snippet powershell example call to `Set-AzVMExtension` included many unnecessary backtick and whitespace characters in the command. These extra characters are most likely from a copy/paste where the powershell command was originally on multiple lines and was copied onto a single line. In powershell adding a backtick at the end of the line indicates that the next line are part of the same command. Since the example in on a single line, the backticks and the extra spaces are not needed.