Updating the right words to avoid confusion
Updating with 'delete' as per the new Azure portal changes. Its no more detach but delete.
Delete word creates a confusion if its deleting the disk completely, so giving a detail description that "the disk is deleted in the VM i.e., detaching the disk", gives more clarification on the action performed. Image also should be changed as per the new portal changes with delete symbol.