Fix an issue with the example code/procedure
So far as I can tell, this issue was first reported here: https://github.com/coreos/afterburn/issues/461
We encountered this same issue, and modifying our procedure in the same way fixed the issue. We have observed an incarnation of 2 appearing in some of our deployments while validating this change.