Grammar and clarity for 2 VMs in FaultDomain note
Some minor grammar changes were made.  In addition, we recommend not following the steps in the note so some language was added to clarify events in a scenario that would lead to the result in question rather than instructions that the reader should follow.  This resolves https://github.com/MicrosoftDocs/azure-docs/issues/86862.