Update SAP account requirements
After converting my SAP user account to a universal ID, the ansible playbook to download the software from started to fail.  Creating a new account which is NOT linked to a SAP Universal ID resolved my issue.  Wanted to call this out as I was stuck for quite a while on this issue.