Update Scheduled Events for New Behaviour
Updated to include the new behaviour in the case that the host is unhealthy.  Also took this chance to add two additional PowerShell samples. There will be another PR following with an update to the Python sample at the bottom, but I didn't want to hold this update on getting that polished

Note - did not update the version and region availability since this change will not impact the version used in the URL. However, this will not be generally available until June