Adding Win Server 2022 SKUs to supported OS images
This PR adds the 4 currently supported Windows Server 2022 SKUs to the supported OS images table.  Not addressed in this PR is an issue with Datacenter and Datacenter Core not yet being supported in the portal and capitalization choices for previous SKUs (this takes the lower case dashed name as it is listed).  This resolves https://github.com/MicrosoftDocs/azure-docs/issues/91588.