Use verbs that apply to all input methods
Similar to https://github.com/MicrosoftDocs/azure-docs/pull/92335

This PR changes instances of "type" to "enter" in order to accommodate non-keyboard users. This is consistent with the Microsoft Style Guide's recommendation for describing interactions with UI.

It also changes bold formatting to italics to be consistent with other instances of user-entered information on this page.