Removing overprovisioning as default statement
"Scale sets currently default to "overprovisioning" VMs."  This has been true in the past but the current experience is quite a mix of enabled or disabled by default depending on the method of deployment.  To avoid confusion, this PR removes the statement declaring overprovisioning as being set by default.  This resolves https://github.com/MicrosoftDocs/azure-docs/issues/93904.