Adding missing resource provider, fixing code
The resource provider "Microsoft.ManagedIdentity" is missing in the Register features section. The user identity would failed if not registered. 

The section Assign permissions for the identity to distribute the images will failed with the message "access to path /myRoleImageCreation.json is denied". Changing the variable $myRoleImageCreationPath = "$env:TEMP\myRoleImageCreation.json" to $myRoleImageCreationPath = "myRoleImageCreation.json" 

Step 4 in the section Create an image will fail with message "a parameter cannot be found that matches parameter name CustomizerName". Changin the parameter "CustomizerName = 'downloadBuildArtifacts' " to "Name = 'downloadBuildArtifacts' "