[Doc-a-thon] Updating custom-script-linux to make it OS agnostic
The small sample script was using apt  which is specific to Ubuntu. Given this is not an Ubuntu specific doc, I changed the script to a sample that could be used on any Linux version (Creating dirs and assigning ownership of an existent user as an example). I also added an actual command to the bash lines, given several of them were only showing directories or paths, not commands