Update generalize.md
I am working as an Azure Cloud Engineer in Microsoft. Recently I received a customer support request where the customer was wondering why only for some of his VMs, he can see DVD E : drive and not for all his VMs. I did some research from my end and got to know that the CD/DVD-ROM drive may be present in some Azure VMs as Azure platform mounts ISO file to a DVD-ROM when a Windows VM is created from generalized image.
For this reason, DVD-ROM must be enabled in guest OS VM. If it is disabled, guest VM is stuck at OOBE even a user believes he did everything correctly.

Users should not attempt to disable DVD-ROM as this will prevent a VM to be provisioned successfully. DVD rom disappears once a VM is deallocated and started again, so this can rarely be an issue.
However, disabling DVD-ROM will yield bad results for image creation job and troubleshooting for this scenario will be very difficult once this issue happens.

When I checked this existing Microsoft document for generalizing a VM before image creation , I found that the above is not mentioned anywhere in it and it can cause serious issues for the customers as well as the troubleshooting engineer.
Hence I have updated the Important notes as well as added a step to enable the DVD-ROM in the guest OS VM.

Please verify from your end and provide your feedback.