Use NVIDIA drivers from Ubuntu
Instead of installing the drivers from NVIDIA, install the drivers present in the Ubuntu archive. Those drivers are the same as the one from NVIDIA but specifically built for the Azure kernel and signed by Canonical. This greatly simplifies the installation process.
We still recommend installing the CUDA toolkit from NVIDIA as it is more up-to-date than the one available in the Ubuntu archive.