Update disks-copy-incremental-snapshot-across-regions.md
There is no info in any public document which states copy feature of full snapshot to another region is not supported. Incase if cx is trying to create full snapshot and copy that snapshot across regions the options are disabled and even region is also disabled. It won't work for powershell/CLI either.  So updating this public document with the restriction - Copy of full snapshots to another region is not supported, only incremental snapshots are supported.