Update disk-encryption-troubleshooting.md
https://packages.ubuntu.com/trusty/linux-azure

Above hyperlink is mentioned in the document which throws error. Please check if we can update with correct one or remove it completely. Below is the line in document:

Under: 
Update the default kernel for Ubuntu 14.04 LTS:
To avoid this error, prior to enabling encryption on the image, update to the Azure tuned kernel 4.15 or later using the following commands: