Adding explicitly requirement to set context of subscription
If we can update this document explicitly defining that subscription for each context should be set in a certain way then it might help in future to have less problems. Had a ticket with Support team and even they couldn't tell me immediately what's the problem. And the problem was that I set context as az cli way and was registering feature via azure powershell and there was no error or hint to tell me that I'm enabling feature for wrong subscription. Ticket is RE: Cant create vm encryption at host - TrackingID#2411200050000262